# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  pagination.Response:
    properties:
      data: {}
      has_more:
        type: boolean
      page:
        type: integer
      per_page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  response.ErrBody:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrBody'
      request_id:
        type: string
      success:
        type: boolean
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  server.adminUserResponse:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      display_name:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      location:
        type: string
      profile_visibility:
        type: string
      role:
        type: string
      status:
        type: string
      username:
        type: string
    type: object
...
  server.authUserResponse:
    properties:
      avatar_url:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  server.changePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  server.publicUserResponse:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      display_name:
        type: string
      id:
        type: string
      location:
        type: string
      profile_visibility:
        type: string
      username:
        type: string
    type: object
...
  server.updateRoleRequest:
    properties:
      role:
        enum:
        - user
        - moderator
        - admin
        - superadmin
        type: string
    required:
    - role
    type: object
  server.updateStatusRequest:
    properties:
      status:
        enum:
        - active
        - suspended
        - banned
        - deactivated
        type: string
    required:
    - status
    type: object
  server.updateUserRequest:
    properties:
      avatar_url:
        maxLength: 2048
        type: string
      bio:
        maxLength: 100
        type: string
      display_name:
        maxLength: 50
        minLength: 3
        type: string
      embed_enabled:
        type: boolean
      location:
        maxLength: 100
        type: string
      profile_visibility:
        enum:
        - public
        - private
        - unlisted
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Standfor API
  title: Standfor API
  version: "1.0"
paths:
  /api/v1/admin/users:
    get:
      description: Returns a paginated list of all users with admin details
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20, max 100)
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pagination.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/server.adminUserResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users (admin)
      tags:
      - admin-users
  /api/v1/admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-deletes a user account
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user (admin)
      tags:
      - admin-users
    get:
      description: Returns full admin details for a user by UUID
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.adminUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID (admin)
      tags:
      - admin-users
  /api/v1/admin/users/{id}/role:
    patch:
      consumes:
      - application/json
      description: Changes a user's role. Only superadmins can assign admin/superadmin
        roles
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: New role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.updateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.adminUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user role (admin)
      tags:
      - admin-users
  /api/v1/admin/users/{id}/status:
    patch:
      consumes:
      - application/json
      description: Changes a user's account status (active, suspended, banned, deactivated)
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.updateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.adminUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user status (admin)
      tags:
      - admin-users
...
  /api/v1/users/{username}:
    get:
      consumes:
      - application/json
      description: Get user by username
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.publicUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get user by username
      tags:
      - users
  /api/v1/users/me:
    delete:
      consumes:
      - application/json
      description: Soft-deletes the authenticated user's account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete current user
      tags:
      - users
    get:
      description: Returns the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.authUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
      - users
    patch:
      consumes:
      - application/json
      description: Updates the authenticated user's profile settings
      parameters:
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.updateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.publicUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - users
  /api/v1/users/me/password:
    post:
      consumes:
      - application/json
      description: Changes the authenticated user's password
      parameters:
      - description: Current and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.changePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - users
swagger: "2.0"